digit-recognition/requirements.txt
tensorflow
streamlit
opencv-python
numpy
pillow
streamlit-drawable-canvas